Implement the paragraph renderer
As part of this commit, we also moved the integration test around so that we
have access to it to guide us later. Looking at the fixture files, we quickly
realized that we'd become very tightly tied to the whitespace formatting of our
renderers, even though multiple subsequent whitespace characters are ignored in
HTML. The fixture file is much easier to read with pretty whitespace though, so
we normalize the whitespace of both the expected and actual values in the test.