Implement the header renderer
We modified our tests to have a test_helper which requires minitest and
sets up our `$LOAD_PATH` so we can stop using `require_relative`
which is basically always wrong. We still have to require `test_helper`
using `require_relative` because we couldn't figure out how to make sure
`test` is always properly on the `$LOAD_PATH` and it wasn't a big enough
deal for us to spend a lot of time on it.

The header renderer turned out to have a lot of edge cases, and took several
iterations to solve, but ultimately we have everything we need. We are not handling
multiline strings because the spec that we need to meet does not specify one way
or the other.